requirements.txt
tensorflow>=2.5.0
numpy>=1.19.2
Pillow>=8.0.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
seaborn>=0.11.0
h5py>=3.1.0